[ut] Custom docker container
There is an already existing docker container for goleft, but it has not been updated in about two years and I don't know who made it or how. For security purposes I figured it may be better to just roll my own. It was also a learning experience, although I am still unsure why Google does not include bzip2 in their golang containers... In any case the [ut] flag is for the quay push. The container was tested locally and found to work.